+++ packages/api-notifications/src/index.ts
@@ -1,0 +1,4 @@
+export * from './getNotificationsSettings'
+export * from './subscribeForWebPushNotifications'
+export * from './unsubscribeFromWebPushNotifications'
+export * from './updateNotificationsSettings'