+++ packages/app-andrew/src/pages/_app.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import App, { AppContext } from 'next/app'
+import Head from 'next/head'
 import '@gtms/styles/scss/global.scss'
 import styles from './styles.scss'
 import { appWithTranslation } from '@gtms/commons/i18n'
@@ -7,6 +8,7 @@ import { Navigation } from 'components/commons/Navigation'
 import { NotificationsActive } from 'components/commons/NotificationsActive'
 import { NavigationDots } from 'components/commons/NavigationDots'
 import { init, initAuthSession } from '@gtms/state-user'
+import { init as initWPN } from '@gtms/state-notification'
 import { LoginWindow } from 'components/commons/LoginWindow'
 // import ReactTooltip from 'react-tooltip' // todo - make sure that we actually needed it, as it creates some issues with SSR
 import { NotificationsSidebar } from 'components/commons/NotificationsSidebar'
@@ -21,10 +23,6 @@ interface GTMSAppProps {
 }
 
 class GTMSApp extends App<GTMSAppProps> {
-  constructor(props: any) {
-    super(props)
-  }
-
   componentDidMount() {
     const { auth } = this.props
 
@@ -35,6 +33,7 @@ class GTMSApp extends App<GTMSAppProps> {
           refreshToken: string
         }
       )
+      initWPN()
     }
   }
 
@@ -42,14 +41,16 @@ class GTMSApp extends App<GTMSAppProps> {
     const { Component, pageProps } = this.props
 
     return (
-      <div className={styles.wrapper}>
-        <Navigation />
-        <div
-          className={styles.bg}
-          style={{ backgroundImage: `url('/images/temp_images/group_bg_3.png')` }} />
-        <NotificationsSidebar />
-        <div className={styles.wrapperPage}>
-          <NavigationDots />
+      <>
+        <Head>
+          <link rel="manifest" href="/manifest.json" />
+        </Head>
+        <div className={styles.wrapper}>
+          <div
+            className={styles.bg}
+            style={{ backgroundImage: `url('/images/temp_images/group_bg_3.png')` }} />
+          <Navigation />
+          <NotificationsSidebar />
           <NotificationsActive />
           <div className={styles.pageContent}>
             <Component {...pageProps} />