+++ packages/state-post/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.5](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.2.4...@gtms/state-post@0.2.5) (2020-06-21)
+
+**Note:** Version bump only for package @gtms/state-post
+
+
+
+
+
+## [0.2.4](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.2.3...@gtms/state-post@0.2.4) (2020-06-19)
+
+**Note:** Version bump only for package @gtms/state-post
+
+
+
+
+
 ## [0.2.3](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-post@0.2.2...@gtms/state-post@0.2.3) (2020-06-18)
 
 **Note:** Version bump only for package @gtms/state-post